Rentify 11 create homepage (#12)
* Remove all comments in css files.

* Fetch items from backend. Create controller for items.

* Add dependencies.

* Add controller for addresses. Add controller advice for handling custom exceptions. Add custom exceptions.

* Remove all unnecessary comments. Remove all unused imports. Format all files.

* Remove unused component.

---------

+++ back-end/src/main/java/com/devminds/rentify/entity/Address.java
@@ -9,7 +9,6 @@ import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
-import lombok.NonNull;
 
 @Data
 @Entity
@@ -17,7 +16,7 @@ import lombok.NonNull;
 public class Address {
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Id
-    private int id;
+    private Long id;
     @Column
     @NotEmpty
     @Size(max = 100)
